Add exec function to kicker machine
